requirements.txt
fastapi
openai
langchain
PyPDF2
pinecone-client
uvicorn
unstructured
-e .